requirements.txt
reflex==0.6.6.post3
python-dotenv==1.0.1
supabase==2.10.0
pillow==11.0.0